+++ +sw_tests/+unit_tests/unittest_spinw_optmagk.m
@@ -11,11 +11,7 @@ classdef unittest_spinw_optmagk < sw_tests.unit_tests.unittest_super
                                  'k', [1; 0; 0]);
     end
     properties (TestParameter)
-        % kbase, rel_tol
-        % For [1 0; 0 1; 0 0], k in each direction can very independently
-        % which is less accurate, increase tol in this case
-        kbase_opts = {{[1; 1; 0], 1e-5}, ...
-                      {[1 0; 0 1; 0 0], 3e-3}};
+        kbase_opts = {[1; 1; 0], [1 0; 0 1; 0 0]};
     end
     methods (TestMethodSetup)
         function setup_chain_model(testCase)            
@@ -85,7 +81,8 @@ classdef unittest_spinw_optmagk < sw_tests.unit_tests.unittest_super
             swobj.addmatrix('label', 'J2', 'value', J2);
             swobj.addcoupling('mat', 'J1', 'bond', 2, 'subidx', 2);
             swobj.addcoupling('mat', 'J2', 'bond', 1);
-            swobj.optmagk('kbase', kbase_opts{1});
+            % Use rng seed for reproducible results
+            swobj.optmagk('kbase', kbase_opts, 'seed', 1);
 
             expected_k = acos(-J2/(2*J1))/(2*pi);
             rel_tol = 1e-5;
@@ -96,7 +93,20 @@ classdef unittest_spinw_optmagk < sw_tests.unit_tests.unittest_super
             expected_mag_str = testCase.default_mag_str;
             expected_mag_str.k = [expected_k; expected_k; 0];
             testCase.verify_val(swobj.mag_str, expected_mag_str, ...
-                                'rel_tol', kbase_opts{2});
+                                'rel_tol', 1e-5);
+        end
+        function test_afm_chain_ndbase_pso_varargin_passed(testCase)
+            testCase.swobj.addmatrix('label', 'J1', 'value', 1);
+            testCase.swobj.addcoupling('mat', 'J1', 'bond', 1);
+            % Verify in default case there is no warning
+            testCase.verifyWarningFree(...
+                @() testCase.swobj.optmagk, ...
+                'pso:convergence');
+            % Test that MaxIter gets passed through to ndbase.pso, triggers
+            % convergence warning
+            testCase.verifyWarning(...
+                @() testCase.swobj.optmagk('MaxIter', 1), ...
+                'pso:convergence');
         end
     end
 end