+++ swfiles/@spinw/optmagk.m
@@ -126,8 +126,8 @@ kones = ones(1,D);
 
 warning('off','sw_readparam:UnreadInput')
 % optimise the energy using particle swarm
-[pOpt0, ~, ~] = ndbase.pso([],@optfun,1/4*kones,'lb',0*kones,'ub',kones,varargin{:},...
-    'TolFun',1e-5,'TolX',1e-5,'MaxIter',1e3);
+[pOpt0, ~, ~] = ndbase.pso([],@optfun,1/4*kones,'lb',0*kones,'ub',kones,...
+    'TolFun',1e-5,'TolX',1e-5,'MaxIter',1e3, varargin{:});
 warning('on','sw_readparam:UnreadInput')
 
 % generate an R-value