+++ CHANGELOG.rst
@@ -67,3 +67,6 @@ Bug Fixes
   Previously this would've caused a crash.
 - Raise error if invalid shape ``kbase`` is provided to ``optmagk``,
   previously it would be silently set to empty
+- Ensure varargin is correctly passed through to ``ndbase.pso`` from
+  ``optmagk``. Previously user provided ``TolFun``, ``TolX`` and
+  ``MaxIter`` would be overwritten by the defaults.